Move test queries to separate files
